requirement.txt
flask==2.3.2
python-dotenv==1.0.0
google-generativeai==0.3.0python 